+++ src/main/java/com/unstoppabledomains/resolution/naming/service/NSConfig.java
@@ -1,0 +1,12 @@
+package com.unstoppabledomains.resolution.naming.service;
+
+import com.unstoppabledomains.config.network.model.Network;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class NSConfig {
+    private Network chainId;
+    private String blockchainProviderUrl;
+}